Change namings that actually make sense
