+++ app/src/androidTest/java/com/softengi/mobcomp/softwareengi_mobile/StepMonitorInstrumentedTest.java
@@ -1,7 +1,6 @@
 package com.softengi.mobcomp.softwareengi_mobile;
 
 import android.support.test.espresso.intent.rule.IntentsTestRule;
-import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 
 import org.junit.Rule;
@@ -24,8 +23,8 @@ import static android.support.test.espresso.matcher.ViewMatchers.withId;
 public class StepMonitorInstrumentedTest {
 
     @Rule
-    public IntentsTestRule<MainActivity> intentsTestRule =
-            new IntentsTestRule<>(MainActivity.class);
+    public IntentsTestRule<LoginActivity> intentsTestRule =
+            new IntentsTestRule<>(LoginActivity.class);
 
     @Test
     public void TEST(){
