+++ app/src/main/java/com/softengi/mobcomp/softwareengi_mobile/Actions/StepAction.java
@@ -1,14 +1,8 @@
-package com.softengi.mobcomp.softwareengi_mobile.Controllers;
+package com.softengi.mobcomp.softwareengi_mobile.Actions;
 
 import android.content.Context;
 import android.widget.Toast;
 
-import com.android.volley.AuthFailureError;
-import com.android.volley.Request;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.JsonObjectRequest;
-import com.softengi.mobcomp.softwareengi_mobile.Utils.RequestQueueSingleton;
 import com.softengi.mobcomp.softwareengi_mobile.Utils.SharedPrefManager;
 import com.softengi.mobcomp.softwareengi_mobile.Utils.VolleyCallback;
 
@@ -17,7 +11,7 @@ import org.json.JSONObject;
 import java.util.HashMap;
 import java.util.Map;
 
-public class StepController {
+public class StepAction {
 
     private static final String url = "steps";
 
@@ -28,7 +22,7 @@ public class StepController {
         map.put("planId", String.valueOf(planId));
         map.put("username", String.valueOf(SharedPrefManager.getInstance(ctx).getUsername()));
 
-        RequestController.createPostRequest(ctx, map, url, new VolleyCallback() {
+        RequestAction.createPostRequest(ctx, map, url, new VolleyCallback() {
             @Override
             public void onSuccessResponse(JSONObject result) {
                 Toast.makeText(ctx, "Saved steps successfully", Toast.LENGTH_SHORT).show();
