+++ app/src/main/res/layout/activity_athlete.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.softengi.mobcomp.softwareengi_mobile.AthleteActivity">
+    tools:context="com.softengi.mobcomp.softwareengi_mobile.MainActivity">
 
     <android.support.design.widget.BottomNavigationView
         android:id="@+id/athlete_nav"
